+++ src/compress.rs
@@ -1,14 +1,14 @@
 use std::convert::TryInto as _;
 use crate::{Image, raw};
 use crate::buf::{OwnedBuf, OutputBuf};
-use crate::common::{Subsamp, Result, Error, get_error};
+use crate::common::{Subsamp, Result, Error};
+use crate::handle::Handle;
 
 /// Compresses raw pixel data into JPEG.
 #[derive(Debug)]
 #[doc(alias = "tjhandle")]
 pub struct Compressor {
-    handle: raw::tjhandle,
-    quality: i32,
+    handle: Handle,
     subsamp: Subsamp,
 }
 
@@ -19,36 +19,45 @@ unsafe impl Send for Compressor {}
 
 impl Compressor {
     /// Create a new compressor instance.
-    #[doc(alias = "tjInitCompress")]
+    #[doc(alias = "tj3Init")]
     pub fn new() -> Result<Compressor> {
-        unsafe {
-            let handle = raw::tjInitCompress();
-            if !handle.is_null() {
-                Ok(Compressor {
-                    handle,
-                    quality: DEFAULT_QUALITY,
-                    subsamp: DEFAULT_SUBSAMP,
-                })
-            } else {
-                Err(get_error(handle))
-            }
-        }
+        let mut handle = Handle::new(raw::TJINIT_TJINIT_COMPRESS)?;
+        handle.set(raw::TJPARAM_TJPARAM_QUALITY, DEFAULT_QUALITY as libc::c_int)?;
+        handle.set(raw::TJPARAM_TJPARAM_SUBSAMP, DEFAULT_SUBSAMP as i32 as libc::c_int)?;
+        Ok(Compressor { handle, subsamp: DEFAULT_SUBSAMP })
     }
 
     /// Set the quality of the compressed JPEG images.
     ///
     /// The quality ranges from 1 (worst) to 100 (best).
-    pub fn set_quality(&mut self, quality: i32) {
-        // TODO: check whether 1 <= quality <= 100 here?
-        self.quality = quality;
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// let mut compressor = turbojpeg::Compressor::new()?;
+    /// compressor.set_quality(95)?;
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    ///
+    /// An error is returned if `quality` is invalid:
+    ///
+    /// ```
+    /// let mut compressor = turbojpeg::Compressor::new()?;
+    /// assert!(compressor.set_quality(120).is_err());
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    #[doc(alias = "TJPARAM_QUALITY")]
+    pub fn set_quality(&mut self, quality: i32) -> Result<()> {
+        self.handle.set(raw::TJPARAM_TJPARAM_QUALITY, quality as libc::c_int)
     }
 
     /// Set the level of chrominance subsampling of the compressed JPEG images.
     ///
     /// Chrominance subsampling can reduce the compressed image size without noticeable loss of
     /// quality (see [`Subsamp`] for more).
-    pub fn set_subsamp(&mut self, subsamp: Subsamp) {
-        self.subsamp = subsamp;
+    #[doc(alias = "TJPARAM_SUBSAMP")]
+    pub fn set_subsamp(&mut self, subsamp: Subsamp) -> Result<()> {
+        self.handle.set(raw::TJPARAM_TJPARAM_SUBSAMP, subsamp as i32 as libc::c_int)
     }
 
     /// Compresses the `image` into `output` buffer.
@@ -79,8 +88,7 @@ impl Compressor {
     ///
     /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
-    #[doc(alias = "tjCompress2")]
-    #[doc(alias = "tjCompress")]
+    #[doc(alias = "tj3Compress8")]
     pub fn compress(&mut self, image: Image<&[u8]>, output: &mut OutputBuf) -> Result<()> {
         image.assert_valid(image.pixels.len());
 
@@ -89,26 +97,27 @@ impl Compressor {
         let pitch = pitch.try_into().map_err(|_| Error::IntegerOverflow("pitch"))?;
         let height = height.try_into().map_err(|_| Error::IntegerOverflow("height"))?;
 
-        let mut output_len = output.len as libc::c_ulong;
+        self.handle.set(
+            raw::TJPARAM_TJPARAM_NOREALLOC,
+            if output.is_owned { 0 } else { 1 } as libc::c_int,
+        )?;
+        let mut output_len = output.len as raw::size_t;
         let res = unsafe {
-            raw::tjCompress2(
-                self.handle,
+            raw::tj3Compress8(
+                self.handle.as_ptr(),
                 pixels.as_ptr(), width, pitch, height, format as libc::c_int,
                 &mut output.ptr, &mut output_len,
-                self.subsamp as libc::c_int, self.quality,
-                if output.is_owned { 0 } else { raw::TJFLAG_NOREALLOC } as libc::c_int,
             )
         };
         output.len = output_len as usize;
-
         if res != 0 {
-            Err(unsafe { get_error(self.handle) })
+            return Err(self.handle.get_error())
         } else if output.ptr.is_null() {
             output.len = 0;
-            Err(Error::Null())
-        } else {
-            Ok(())
+            return Err(Error::Null)
         }
+
+        Ok(())
     }
 
     /// Compresses the `image` into an owned buffer.
@@ -148,15 +157,9 @@ impl Compressor {
     /// subsampling (see [`set_subsamp()`](Compressor::set_subsamp)).
     ///
     /// You can also use [`compressed_buf_len()`] directly.
-    #[doc(alias = "tjBufSize")]
+    #[doc(alias = "tj3JPEGBufSize")]
     pub fn buf_len(&self, width: usize, height: usize) -> Result<usize> {
-        super::compressed_buf_len(width, height, self.subsamp)
-    }
-}
-
-impl Drop for Compressor {
-    fn drop(&mut self) {
-        unsafe { raw::tjDestroy(self.handle); }
+        compressed_buf_len(width, height, self.subsamp)
     }
 }
 
@@ -182,8 +185,8 @@ impl Drop for Compressor {
 /// ```
 pub fn compress(image: Image<&[u8]>, quality: i32, subsamp: Subsamp) -> Result<OwnedBuf> {
     let mut compressor = Compressor::new()?;
-    compressor.set_quality(quality);
-    compressor.set_subsamp(subsamp);
+    compressor.set_quality(quality)?;
+    compressor.set_subsamp(subsamp)?;
     compressor.compress_to_owned(image)
 }
 
@@ -193,11 +196,11 @@ pub fn compress(image: Image<&[u8]>, quality: i32, subsamp: Subsamp) -> Result<O
 ///
 /// Returns an error on integer overflow. You can just `.unwrap()` the result if you don't care
 /// about this edge case.
-#[doc(alias = "tjBufSize")]
+#[doc(alias = "tj3JPEGBufSize")]
 pub fn compressed_buf_len(width: usize, height: usize, subsamp: Subsamp) -> Result<usize> {
     let width = width.try_into().map_err(|_| Error::IntegerOverflow("width"))?;
     let height = height.try_into().map_err(|_| Error::IntegerOverflow("height"))?;
-    let len = unsafe { raw::tjBufSize(width, height, subsamp as libc::c_int) };
+    let len = unsafe { raw::tj3JPEGBufSize(width, height, subsamp as libc::c_int) };
     let len = len.try_into().map_err(|_| Error::IntegerOverflow("buf len"))?;
     Ok(len)
 }